requirements.txt
Flask==3.0.0
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
google-auth==2.27.0
python-dotenv==1.0.0
yt-dlp==2025.1.26
openai==1.56.1
httpx==0.27.2
langdetect==1.0.9